Lab 12 final/requirements.txt
pytest==7.3.1
Pillow==10.0.0
pyinstaller==6.0.0 